+++ service/pipelines/config_test.go
@@ -9,15 +9,18 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 
 	"go.opentelemetry.io/collector/component"
+	"go.opentelemetry.io/collector/featuregate"
 	"go.opentelemetry.io/collector/pipeline"
+	"go.opentelemetry.io/collector/pipeline/pipelineprofiles"
 )
 
 func TestConfigValidate(t *testing.T) {
 	var testCases = []struct {
 		name     string // test case name (also file name containing config yaml)
-		cfgFn    func() Config
+		cfgFn    func(*testing.T) Config
 		expected error
 	}{
 		{
@@ -27,8 +30,8 @@ func TestConfigValidate(t *testing.T) {
 		},
 		{
 			name: "duplicate-processor-reference",
-			cfgFn: func() Config {
-				cfg := generateConfig()
+			cfgFn: func(*testing.T) Config {
+				cfg := generateConfig(t)
 				pipe := cfg[pipeline.NewID(pipeline.SignalTraces)]
 				pipe.Processors = append(pipe.Processors, pipe.Processors...)
 				return cfg
@@ -37,8 +40,8 @@ func TestConfigValidate(t *testing.T) {
 		},
 		{
 			name: "missing-pipeline-receivers",
-			cfgFn: func() Config {
-				cfg := generateConfig()
+			cfgFn: func(*testing.T) Config {
+				cfg := generateConfig(t)
 				cfg[pipeline.NewID(pipeline.SignalTraces)].Receivers = nil
 				return cfg
 			},
@@ -46,8 +49,8 @@ func TestConfigValidate(t *testing.T) {
 		},
 		{
 			name: "missing-pipeline-exporters",
-			cfgFn: func() Config {
-				cfg := generateConfig()
+			cfgFn: func(*testing.T) Config {
+				cfg := generateConfig(t)
 				cfg[pipeline.NewID(pipeline.SignalTraces)].Exporters = nil
 				return cfg
 			},
@@ -55,15 +58,15 @@ func TestConfigValidate(t *testing.T) {
 		},
 		{
 			name: "missing-pipelines",
-			cfgFn: func() Config {
+			cfgFn: func(*testing.T) Config {
 				return nil
 			},
 			expected: errMissingServicePipelines,
 		},
 		{
 			name: "invalid-service-pipeline-type",
-			cfgFn: func() Config {
-				cfg := generateConfig()
+			cfgFn: func(*testing.T) Config {
+				cfg := generateConfig(t)
 				cfg[pipeline.MustNewID("wrongtype")] = &PipelineConfig{
 					Receivers:  []component.ID{component.MustNewID("nop")},
 					Processors: []component.ID{component.MustNewID("nop")},
@@ -73,17 +76,50 @@ func TestConfigValidate(t *testing.T) {
 			},
 			expected: errors.New(`pipeline "wrongtype": unknown signal "wrongtype"`),
 		},
+		{
+			name: "disabled-featuregate-profiles",
+			cfgFn: func(*testing.T) Config {
+				cfg := generateConfig(t)
+				cfg[pipeline.NewID(pipelineprofiles.SignalProfiles)] = &PipelineConfig{
+					Receivers:  []component.ID{component.MustNewID("nop")},
+					Processors: []component.ID{component.MustNewID("nop")},
+					Exporters:  []component.ID{component.MustNewID("nop")},
+				}
+				return cfg
+			},
+			expected: errors.New(`pipeline "profiles": profiling signal support is at alpha level, gated under the "service.profilesSupport" feature gate`),
+		},
+		{
+			name: "enabled-featuregate-profiles",
+			cfgFn: func(t *testing.T) Config {
+				require.NoError(t, featuregate.GlobalRegistry().Set(serviceProfileSupportGateID, true))
+
+				cfg := generateConfig(t)
+				cfg[pipeline.NewID(pipelineprofiles.SignalProfiles)] = &PipelineConfig{
+					Receivers:  []component.ID{component.MustNewID("nop")},
+					Processors: []component.ID{component.MustNewID("nop")},
+					Exporters:  []component.ID{component.MustNewID("nop")},
+				}
+				return cfg
+			},
+			expected: nil,
+		},
 	}
 
 	for _, tt := range testCases {
 		t.Run(tt.name, func(t *testing.T) {
-			cfg := tt.cfgFn()
+			cfg := tt.cfgFn(t)
 			assert.Equal(t, tt.expected, cfg.Validate())
+
+			// Clean up the profiles support gate, which may have been enabled in `cfgFn`.
+			require.NoError(t, featuregate.GlobalRegistry().Set(serviceProfileSupportGateID, false))
 		})
 	}
 }
 
-func generateConfig() Config {
+func generateConfig(t *testing.T) Config {
+	t.Helper()
+
 	return map[pipeline.ID]*PipelineConfig{
 		pipeline.NewID(pipeline.SignalTraces): {
 			Receivers:  []component.ID{component.MustNewID("nop")},